Refactored event generation
