+++ requirements/EVENT_SERVICES_AURORA_0002/send-event-notifications/work/eventgenerator3/C/src/eventgenerator3.c
@@ -8,46 +8,9 @@
     !! file. The up-to-date signatures can be found in the .ads file.   !!
 */
 #include "eventgenerator3.h"
-#include <string.h>
-#include <stdlib.h>
-#include <time.h>
-
-#define BIT_MASK 0xFF
+#include "event_helper.h"
 
 static int counter;
-
-asn1SccT_EventMessage eventgenerator3_buildEvent(asn1SccT_UInt32 id, asn1SccT_EventMessage_eventType eventType)
-{
-    char message[] = "This is message from EventGenerator3";
-    uint64_t seconds = 0;
-    uint64_t nanoseconds = 0;
-    struct timespec ts;
-    int return_code = timespec_get(&ts, TIME_UTC);
-
-    if(return_code != 0)
-    {
-        seconds = (uint64_t)ts.tv_sec;
-        nanoseconds = (uint64_t)ts.tv_nsec;
-    }
-
-    asn1SccT_CUCTimestamp timestamp;
-    timestamp.c1 = (seconds >> 24) & BIT_MASK;
-    timestamp.c2 = (seconds >> 16) & BIT_MASK;
-    timestamp.c3 = (seconds >> 8) & BIT_MASK;
-    timestamp.c4 = seconds & BIT_MASK;
-    timestamp.f1 = (nanoseconds >> 24) & BIT_MASK;
-    timestamp.f2 = (nanoseconds >> 16) & BIT_MASK;
-    timestamp.f3 = (nanoseconds >> 8) & BIT_MASK;
-
-    asn1SccT_EventMessage event;
-    event.id = id;
-    event.eventType = eventType;
-    strncpy(event.message.arr, message, sizeof(event.message.arr) - 1);
-    event.message.arr[sizeof (event.message.arr) - 1] = '\0';
-    event.timestamp = timestamp;
-
-    return event;
-}
 
 void eventgenerator3_startup(void)
 {
@@ -60,12 +23,14 @@ void eventgenerator3_PI_trigger3(void)
 
     if(counter == 0)
     {
-        event = eventgenerator3_buildEvent(3, T_EventMessage_eventType_highSeverity);
+        event = build_event(3, T_EventMessage_eventType_highSeverity,
+                            "This is message from EventGenerator3");
         counter++;
     }
     else
     {
-        event = eventgenerator3_buildEvent(4, T_EventMessage_eventType_informative);
+        event = build_event(4, T_EventMessage_eventType_informative,
+                            "This is message from EventGenerator3");
         counter = 0;
     }
